Format

+++ src/helper/deleteImageHelper.ts
@@ -2,10 +2,9 @@ import fs from 'fs/promises'
 
 export const deleteImage = async (imagePath: string) => {
   try {
+    if (imagePath == 'public/images/users/default.png') return
     await fs.unlink(imagePath)
-    console.log('Image is deleted from server')
   } catch (error) {
-    console.log(error)
     throw error
   }
 }
